# app/static/swagger.json
{
...
  "info": {
    "description": "A RESTful API for a sports team and player matching platform.",
    "title": "Sport Matching API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/avatar": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "fileUrl": {
                    "type": "string"
                  }
                },
                "required": [
                  "fileUrl"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Avatar updated successfully."
          },
          "400": {
            "description": "Missing fileUrl."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Upload an avatar for a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "responses": {
          "200": {
            "description": "The current user's profile."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get current user",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/users/{user_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "A single user's profile."
          },
          "404": {
            "description": "User not found."
          }
        },
        "summary": "Get a user by ID",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "type": "integer"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "city": {
                    "type": "string"
                  },
                  "nickname": {
                    "type": "string"
                  },
                  "sports": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User updated successfully."
          },
          "400": {
            "description": "No data provided."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "Users"
        ]
      }
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}